Final Commit for DebtCollector project. Fixed some issues with the RecyclerView.
Issues were as follows:

- Scaling and positioning issues across different screen sizes
- Positioning issues on orientation change
-Text size issues across different screen sizes
-API compatibility below API 22

These issues were resolved by setting padding relative to screen size. Position issue was a result of using the wrong variable to adjust the position of the GraphView (now corrected to layout_gravity)

Positioning issue of orientation change was resovlved by setting position based on first/last visible position when configuration changes occur.

Text size issue was resolved by reducing the size of the default message text. This was the only line of text causing an issue as the length meant it would spill onto a second line if the screen size was small. Because of this all other text sizes remain the same.

Rather than rewrite the code for backwards compatibility beyond API 22 I've decided to set the minimum API to 22. According to the latest numbers fewer than 15% of android users are using APIs below this level, a number which I'm certain will soon decrease to a negligible amount.

As a result of these changes the app is now suitable for use on a wide range of devices.